Begin adding geocoder client
As a first pass, add hardcode a query string and retrieve a response
from the geocoder api. Insert the relevant parts into a slice.

In order to bein calling the api, send a get request to the search endpoint
which asks for information about the statue of libery. Create a struct with
the longitude and latitude and the display name of the api responses.

Next steps: Accept user input instead of the statue of liberty string.
Make this callable and pass the slice back to the caller.
back to the caller