Make geocoder a package
Add a folder called geocoder and add the geocoder.go file in there.

Add an initialiser for the geocoderClient so that it can be created and
used in other packages. Add functions to the geocoder which take the address
as a string argument and retun an array of places to the caller.

Example usage:
```
import "whats-the-weather/main/geocoder"

geoClient := geocoder.NewGeocoderClient()
resp, err := geoClient.FindCoordinates("Statue of Liberty")
```
	renamed:    geocoder.go -> geocoder/geocoder.go