Add a test for geocode.
I want to test that the geocode function returns what is expected
given the response from the geocode api.

In order to mock the response of the geocode api using the http.Server
test mock, it is necessary to refactor the geocode method so that it
accepts the url to GET from as an argument. If it accepts it as an argument
then we can mock it