Add Geocode client
In order to be able to mock calls to a geocode client and to test that
the application moulds the right response from the geocode response, we
need to create a geocode interface. Then, any goecode package can be wrapped
in the geocode client.

Imagine we want to switch to use google maps as our geocode provider,
in that case we'll just have to make sure that it implements the
fullURL and getPlace methods.

Note: At the moment, geocoder.maps is the default geocoder. Notice that
it is made a default by allowing nil to be passed as an arg to the geocoder
client. I'm not sure if this is best practice! But it will do for now as the main
reason for this restructure was so that we can mock in tests