Add a cache
Why?
We need to cache the responses from the yr.no and geocoder api

How?
Generate a uuid and store the info in a file with the uuid as a name. Store the
key and the uuid in a map using bitcask. If a key is requested, look up the uuid and
read from the correct file.

If we open a connection, we need to also close it. This is a bit tricky when you're writing
a packaged so I have decided to open and close a db connection in every request. I don't expect
that this will have heavy use (eg hundreds of reads/writes per s) so I don't expct that
this will slow down and it is the safest way to make sure that the connection is definitely closed