Use the cache in the location command
Why?
The co ordinates of a location will never change, this means that we can
just cache them forever and use the cached ones. This prevents us from
having to call the geocoder api

How?
Use the requested location name as the cache key. Check the cache before
querying the geocoder api for the coordinates of a location.