Add tests for caching module
Test creating a cache, adding to it and getting from it.

In order to be able to test this using temporary test folder, we need to
change the cache module so that it accepts the db location and cache
folder location as arguments. This means that we can use different ones
for testing and they'll be cleaned up properly