Add a caching implementatin for the forecast
Why?
When the yr.no api returns a response, it tells you the time that the
forecast it gives you will expire. If the user requests forcast for
thesame location within this timeframe, there is no need to get it from
yr.no. Instead we can cache it with an expiry time.

How?
Create a struct which holds the expriy time and the forecast. We can
serialize and deserialize that json into the existing cache.
The forecast cache should be in charge of deciding whether
an object in the cache has expired. If the client asks the forecast
cache for an object, and there is an object in the cache but it is
expired, the cache should not return the object.