Add caching with expiry to the cache package
Why?
Each request to yr.no comes with an expiry time. We should not ask for
the same forecast again if the first one is not expired.

How?
Extend the cache package to use PutWithTTL to the database. This will
save an object. When we ask to retrieve that object, if it is expired
it will throw a KeyExpired error.

In order to allow this to happen, we need to change the cache
implementation to pass through the errors which are returned from
bitcask.

Currently all of the logic is inside the command function. Ideally, this
could be split out into a package.