Delete the cached data
The cache writes data to disk, if there is a lot of data the file size
could get very big. This command allows the user to delete the cache.

Deleting the cache in this way also deletes the database. It relies on
the fact that we find or create a database when we write to it in the
cli.