Make cache location fixed
Why?
When we run the CLI, it shouldn't matter what folder on a machine it is
run from, the datastore should always be in the same place.

How?
When we create a data store using the cahce module, always create it in
the home directory of the machine in a hidden folder.