Add debug logs
Why?
When I run this cli locally, the home directory is selected correctly.
However, when I install ot using brew the cache is not being written to
the right file.

This commit adds a debug flag to the root command and some structured
logging which will print verbose debug logs to the console to help with
debugging"